requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
PyMuPDF
wikipedia
langchain_text_splitters
langchain-openai
langchain_ollama
chromadb
faiss-cpu
langchain_chroma
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette
transformers
langchain_huggingface
sentence-transformers